+++ meta_package_manager/managers/apt.py
@@ -232,6 +232,7 @@ class APT(PackageManager):
             ► apt install --quiet bat
 
         """
+        super().install(package_id)
         return self.run_cli("install", self.global_args, package_id)
 
     @property