+++ meta_package_manager/managers/mas.py
@@ -132,6 +132,7 @@ class MAS(PackageManager):
             ► mas install 945397020
 
         """
+        super().install(package_id)
         return self.run_cli(self.global_args, "install", package_id)
 
     @property