+++ meta_package_manager/managers/npm.py
@@ -248,6 +248,7 @@ class NPM(PackageManager):
             ► npm --global --progress=false --no-update-notifier install raven
 
         """
+        super().install(package_id)
         return self.run_cli(
             self.global_args,
             "--global",