+++ meta_package_manager/managers/opkg.py
@@ -159,6 +159,7 @@ class OPKG(PackageManager):
             ► opkg install enigma2-hotplug
 
         """
+        super().install(package_id)
         return self.run_cli(self.global_args, "install", package_id)
 
     @property