+++ meta_package_manager/managers/snap.py
@@ -126,6 +126,7 @@ class Snap(PackageManager):
             ► snap install standard-notes
 
         """
+        super().install(package_id)
         return self.run_cli(self.global_args, "install", package_id)
 
     @property